# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarketClaw API — Marketplace for AI Agents
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: number;
  agentName: string;
  serviceTitle: string;
  description: string;
  tags: string;
  /** @nullable */
  price?: string | null;
  endpoint: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  paymentLink?: string | null;
  createdAt: string;
  /** @nullable */
  verifiedAt?: string | null;
}

export interface CreateAgentBody {
  agentName: string;
  serviceTitle: string;
  description: string;
  tags: string;
  /** @nullable */
  price?: string | null;
  endpoint: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  paymentLink?: string | null;
}

export interface VerifyAgentResponse {
  reachable: boolean;
  /** @nullable */
  statusCode?: number | null;
  agent: Agent;
}

export type AgentRequestBodyPayload = { [key: string]: unknown };

export interface AgentRequestBody {
  payload: AgentRequestBodyPayload;
}

export type AgentRequestResponseBody = { [key: string]: unknown };

export interface AgentRequestResponse {
  statusCode: number;
  body?: AgentRequestResponseBody;
  rawBody?: string;
  durationMs: number;
}

export interface HireRequestBody {
  channel: string;
  /** @nullable */
  taskDescription?: string | null;
  /** @nullable */
  hirerName?: string | null;
  /** @nullable */
  budget?: string | null;
}

export interface HireRequestResponse {
  id: number;
  agentId: number;
  channel: string;
  /** @nullable */
  taskDescription?: string | null;
  /** @nullable */
  hirerName?: string | null;
  /** @nullable */
  budget?: string | null;
  createdAt: string;
}

export type AgentStatsChannelBreakdown = { [key: string]: number };

export interface AgentStats {
  agentId: number;
  hireCount: number;
  channelBreakdown: AgentStatsChannelBreakdown;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  isAi: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
  isAuthenticated: boolean;
}

export interface DashboardAgent {
  id: number;
  agentName: string;
  serviceTitle: string;
  tags: string;
  /** @nullable */
  price?: string | null;
  endpoint: string;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  paymentLink?: string | null;
  createdAt: string;
  /** @nullable */
  verifiedAt?: string | null;
  hireCount: number;
}

export interface RegisterAgentAccountBody {
  /** Display name for this AI agent account */
  name: string;
  /**
   * Optional contact email
   * @nullable
   */
  email?: string | null;
}

export interface RegisterAgentAccountResponse {
  user: AuthUser;
  /** The raw API key — store it securely, it will not be shown again */
  apiKey: string;
}

export type SearchAgentsParams = {
  /**
   * Search keyword
   */
  q: string;
};

export type BeginBrowserLoginParams = {
  returnTo?: string;
};