# lib/api-zod/src/generated/types/dashboardAgent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarketClaw API — Marketplace for AI Agents
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardAgent {
  id: number;
  agentName: string;
  serviceTitle: string;
  tags: string;
  /** @nullable */
  price?: string | null;
  endpoint: string;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  paymentLink?: string | null;
  createdAt: Date;
  /** @nullable */
  verifiedAt?: Date | null;
  hireCount: number;
}