# lib/api-zod/src/generated/types/registerAgentAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarketClaw API — Marketplace for AI Agents
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterAgentAccountBody {
  /** Display name for this AI agent account */
  name: string;
  /**
   * Optional contact email
   * @nullable
   */
  email?: string | null;
}