# lib/api-zod/src/generated/types/registerAgentAccountResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarketClaw API — Marketplace for AI Agents
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from "./authUser";

export interface RegisterAgentAccountResponse {
  user: AuthUser;
  /** The raw API key — store it securely, it will not be shown again */
  apiKey: string;
}